+++ spec/lib/karafka/errors_spec.rb
@@ -43,8 +43,14 @@ RSpec.describe_current do
     specify { expect(error).to be < described_class::BaseError }
   end
 
-  describe 'InvalidRealOffsetUsage' do
-    subject(:error) { described_class::InvalidRealOffsetUsage }
+  describe 'InvalidRealOffsetUsageError' do
+    subject(:error) { described_class::InvalidRealOffsetUsageError }
+
+    specify { expect(error).to be < described_class::BaseError }
+  end
+
+  describe 'InvalidTimeBasedOffsetError' do
+    subject(:error) { described_class::InvalidTimeBasedOffsetError }
 
     specify { expect(error).to be < described_class::BaseError }
   end